+++ changelog_gen/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.3.0"
+__version__ = "0.6.0"
 VERSION = __version__